Permit system memory to be up to 4194303K (4GB less 1K)

+++ src/heap.h
@@ -32,9 +32,9 @@
 
 extern boolean init_heap(void);
 extern void release_heap(void);
-extern boolean init_workspace(uint32);
+extern boolean init_workspace(size_t);
 extern void release_workspace(void);
-extern void *allocmem(int32, boolean);
+extern void *allocmem(size_t, boolean);
 extern boolean returnable(void *, int32);
 extern void freemem(void *, int32);
 extern void clear_heap(void);
